+++ scripts/download-release-assets.js
@@ -1,5 +1,6 @@
 const {downloadReleaseAssets, getLatestTag} = require('./github-client');
 const fs = require('fs');
+const path = require('path');
 
 async function main() {
   const tag = await getLatestTag();
@@ -7,6 +8,7 @@ async function main() {
   // be able to do it's job properly.
   const topLevelDirectory = './artifacts';
   const subDirectoryForTarball = `${topLevelDirectory}/coveo-${tag}`;
+  const binariesMatcher = /^coveo(_|-)(?<_version>v?\d+\.\d+\.\d+(-\d+)?)(?<longExt>.*\.(exe|deb|pkg))$/;
 
   if (!fs.existsSync(topLevelDirectory)) {
     fs.mkdirSync(topLevelDirectory);
@@ -24,6 +26,19 @@ async function main() {
       console.info(assetName, `--> ${topLevelDirectory}`);
       return topLevelDirectory;
     }
+  });
+
+  const files = fs.readdirSync(topLevelDirectory, {withFileTypes: true});
+  files.forEach((file) => {
+    const match = binariesMatcher.exec(file.name);
+    if (!match) {
+      return;
+    }
+    const destName = `coveo-latest${match.groups.longExt}`;
+    fs.copyFileSync(
+      path.resolve(topLevelDirectory, file.name),
+      path.resolve(topLevelDirectory, destName)
+    );
   });
 }
